requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.0
aiofiles==23.2.1
python-dotenv==1.0.0
email-validator==2.1.0
psycopg2-binary==2.9.9
alembic==1.12.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
fastapi-admin==1.0.4
jinja2==3.1.2
asyncpg==0.28.0
tortoise-orm==0.20.0